requirements.txt
Django==2.1.2
django-debug-toolbar==1.10.1
django-filter==2.0.0
django-model-utils==3.1.2
djangorestframework==3.9.0
hiredis==0.2.0
psycopg2==2.7.5
pytz==2018.5
redis==2.10.6
sqlparse==0.2.4
